+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 <!-- markdownlint-disable no-trailing-punctuation -->
 
+## next
+
+TODO: Date
+
+* Use `#let ident` to declare `ident`.
+
 ## 0.0.5
 
 2020-11-14